requirements.txt
Flask==3.0.3
lxml==5.1.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
flask-login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
deep-translator==1.11.4